+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.3
+
+- fix multiple selections not working correctly
+- fix duplicate with new line indentations, remove config `formatAfterDuplication` which requires a formatter